Do not use deprecated APIs and adjust requirements to Xfce 4.10
Raise the minimum requirements of Xfce components to 4.10 which is the
currently supported version.
Raise the minimum requirements of GTK/glib to match those of Xfce 4.10.
Raise the minimum required gstreamer version to 0.10.23, it was released in
2009 and will allow us to use GST_MIXER_FLAG_HAS_WHITELIST in the mixer.
Replace the usage of GTK/glib API parts which have been deprecated in the
minimum required GTK version.