[sndio] make sure to save a non-zero volume when setting mute (fixes #5)
when the volume slider is lowered to 0, gst_mixer_sndio_set_mute() is called
with mute=1, and thus the saved value is 0. When trying to increment the volume
later on, gst_mixer_sndio_set_mute() is called with mute=0, we then restore the
saved volume .. to 0. Thus, the volume slider stays stuck on 0. FAIL!